chore: update extern to match integral MHz

+++ runtime/include/runtime.h
@@ -28,7 +28,7 @@ extern FILE *runtime_sandbox_perf_log;
  * Assumption: All cores are the same speed
  * See runtime_get_processor_speed_MHz for further details
  */
-extern float runtime_processor_speed_MHz;
+extern uint32_t runtime_processor_speed_MHz;
 
 /* Count of worker threads and array of their pthread identifiers */
 extern pthread_t runtime_worker_threads[];
